+++ src/main/java/ladder/ladder/LadderGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
