+++ src/main/java/ladder/ladder/Position.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.ladder;
 
 import java.util.Objects;
