+++ src/main/java/ladder/user/UserName.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.user;
 
 import java.util.Objects;
