+++ src/main/java/ladder/view/ResultView.java
@@ -1,6 +1,13 @@
-package ladder;
+package ladder.view;
 
 import java.util.List;
+import ladder.result.Award;
+import ladder.result.Awards;
+import ladder.ladder.Ladder;
+import ladder.result.LadderResult;
+import ladder.ladder.Row;
+import ladder.user.UserName;
+import ladder.user.UserNames;
 
 public class ResultView {
