+++ src/test/java/ladder/ladder/PositionTest.java
@@ -1,8 +1,9 @@
-package ladder;
+package ladder.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ladder.ladder.Position;
 import org.junit.jupiter.api.Test;
 
 class PositionTest {
