+++ src/test/java/ladder/ladder/RowGeneratorTest.java
@@ -1,8 +1,10 @@
-package ladder;
+package ladder.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.IntStream;
+import ladder.ladder.Row;
+import ladder.ladder.RowGenerator;
 import org.junit.jupiter.api.Test;
 
 class RowGeneratorTest {
