+++ src/test/java/ladder/ladder/RowTest.java
@@ -1,8 +1,9 @@
-package ladder;
+package ladder.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import ladder.ladder.Row;
 import org.junit.jupiter.api.Test;
 
 class RowTest {
